Reset conformance
Currently the conformance code for writing isn't actually running. So
I'll just remove it here.
