+++ rust/examples/common/conformance_writer_spec.rs
@@ -20,7 +20,7 @@ pub struct Record {
 }
 
 impl Record {
-    pub fn get_field(&self, name: &str) -> &Value {
+    pub fn get_field(self: &Self, name: &str) -> &Value {
         return &self
             .fields
             .iter()
@@ -29,19 +29,19 @@ impl Record {
             .1;
     }
 
-    pub fn get_field_data(&self, name: &str) -> Vec<u8> {
+    pub fn get_field_data(self: &Self, name: &str) -> Vec<u8> {
         let data: Vec<u8> = self
             .get_field(name)
             .as_array()
             .unwrap_or_else(|| panic!("Invalid: {}", name))
-            .iter()
+            .into_iter()
             .filter_map(|v| v.as_u64())
             .filter_map(|n| u8::try_from(n).ok())
             .collect();
-        data
+        return data;
     }
 
-    pub fn get_field_meta(&self, name: &str) -> BTreeMap<String, String> {
+    pub fn get_field_meta(self: &Self, name: &str) -> BTreeMap<String, String> {
         let data = self
             .get_field(name)
             .as_object()
@@ -50,17 +50,17 @@ impl Record {
         for (key, value) in data.iter() {
             result.insert(key.to_string(), value.as_str().unwrap().to_string());
         }
-        result
+        return result;
     }
 
-    pub fn get_field_str(&self, name: &str) -> &str {
+    pub fn get_field_str(self: &Self, name: &str) -> &str {
         return self
             .get_field(name)
             .as_str()
             .unwrap_or_else(|| panic!("Invalid: {}", name));
     }
 
-    pub fn get_field_u16(&self, name: &str) -> u16 {
+    pub fn get_field_u16(self: &Self, name: &str) -> u16 {
         return self
             .get_field(name)
             .as_str()
@@ -68,7 +68,7 @@ impl Record {
             .unwrap_or_else(|| panic!("Invalid: {}", name));
     }
 
-    pub fn get_field_u32(&self, name: &str) -> u32 {
+    pub fn get_field_u32(self: &Self, name: &str) -> u32 {
         return self
             .get_field(name)
             .as_str()
@@ -76,7 +76,7 @@ impl Record {
             .unwrap_or_else(|| panic!("Invalid: {}", name));
     }
 
-    pub fn get_field_u64(&self, name: &str) -> u64 {
+    pub fn get_field_u64(self: &Self, name: &str) -> u64 {
         return self
             .get_field(name)
             .as_str()
